pyproject.toml
[tool.poetry]
name = "stacks-data"
version = "0.1.0"
description = "A suite of utilities to support data engineering workloads within an Ensono Stacks data platform."
authors = [
    "Andy Durkan <andy.durkan@ensono.com>",
    "Paulina Motyl <paulina.motyl@ensono.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
pyspark = "^3.4.0"
jsonschema = "^4.17.3"
azure-identity = "^1.13.0"
azure-mgmt-datafactory = "^3.1.0"
azure-storage-file-datalake = "^12.11.0"
click-loglevel = "^0.4.0.post1"
colorlog = "^6.7.0"
polling2 = "^0.5.0"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"
great-expectations = "^0.17.11"
pydantic = "^1.10.12"
delta-spark = "^2.4.0"
behave = "^1.2.6"

[tool.poetry.scripts]
datastacks = "datastacks.cli.datastacks_cli:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.1.0"
pre-commit = "^3.2.0"
flake8 = "^6.0.0"
flake8-docstrings = "^1.7.0"
pydocstyle = "^6.3.0"
tomli = "^2.0.1"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"

[tool.black]
    line-length = 120
    target-version = ['py310']
    exclude = '''
    /(
        \.git
      | \.venv
    )/
    '''

[tool.flake8]
max-line-length = "120"
exclude = ".venv,.git"

[tool.pydocstyle]
convention = "google"
add-ignore = "D100,D101,D104"
match = "(?!test_|conftest).*\\.py"
match_dir = "^(?!(tests)).*"
